+++ tradingagents/graph/trading_graph.py
@@ -122,6 +122,7 @@ class TradingAgentsGraph:
             self.deep_thinking_llm,
             self.tool_nodes,
             self.conditional_logic,
+            self.config.get("enable_execution_levels", False),
         )
 
         self.propagator = Propagator()